Render Components In Style with CSS
