+++ client/App.css
@@ -1,0 +1,17 @@
+.product_detail {
+  width: 100%;
+  margin: 0;
+  padding: 0;
+  font-size: 0.75em;
+  background-color: #fff;
+}
+
+.product_detail_header {
+  position: relative;
+  z-index: 4;
+  background: #ffffff;
+  margin: 0 auto;
+  padding: 14px 0 16px;
+  min-width: 1024px;
+  max-width: 1132px;
+}
