+++ client/App.js
@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
 import { Link, Route } from 'react-router-dom';
-
-import Event from './container/EventDetail';
 import asyncComponent from './hoc/asyncComponent';
+
+import classes from './App.css';
+
+import ProductDetail from './container/ProductDetail/ProductDetail';
 
 const AsyncPizza = asyncComponent(() => {
   return import('./container/Pizza.js');
@@ -11,12 +13,12 @@ const AsyncPizza = asyncComponent(() => {
 class App extends Component {
   render() {
     return (
-      <div>
+      <div className={classes.product_detail}>
         <div>
-          <Link to={'/'}>Event</Link> | <Link to="/pizza">Pizza</Link>
+          <Link to={'/'}>Detail</Link> | <Link to="/pizza">Pizza</Link>
         </div>
-        <div>
-          <Route path="/" exact component={Event} />
+        <div className={classes.product_detail_header}>
+          <Route path="/" exact component={ProductDetail} />
           <Route path="/pizza" component={AsyncPizza} />
         </div>
       </div>
