--- client/components/EventDesc/EventDesc.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import classes from './EventDesc.css';
-
-const eventDescription = (props) => {
-  return <div />;
-};
-
-export default eventDescription;
