--- client/components/EventImage/EventImage.css
@@ -1,11 +1,0 @@
-.EventImage {
-  width: 80%;
-  height: 300px;
-  margin: 20px auto;
-  text-align: center;
-}
-
-.EventImg {
-  max-width: 100%;
-  max-height: 100%;
-}
