+++ client/components/Highlights/Highlights.css
@@ -1,0 +1,4 @@
+.product_highlight_module {
+  flex-basis: 50%;
+  padding: 11px 0 0 28px;
+}
