+++ client/components/PhotoModule/PhotoModule.css
@@ -1,0 +1,61 @@
+/* .PhotoImage {
+  width: 80%;
+  height: 300px;
+  margin: 20px auto;
+  text-align: center;
+}
+
+.PhotoImg {
+  max-width: 100%;
+  max-height: 100%;
+} */
+
+.photo_module {
+  flex-basis: 552px;
+  flex-shrink: 0;
+  background-color: transparent;
+  padding: 0;
+  border-right: none;
+  position: relative;
+}
+
+.perma {
+  position: relative;
+  z-index: 0;
+  width: 552px;
+  height: 364px;
+  line-height: 334px;
+  overflow: hidden;
+  margin-left: auto;
+  margin-right: auto;
+  -webkit-transition: width 200ms linear;
+  -moz-transition: width 200ms linear;
+  -ms-transition: width 200ms linear;
+  -o-transition: width 200ms linear;
+  transition: width 200ms linear;
+}
+
+.carouselPhoto {
+  width: 200%;
+  height: 100%;
+  margin-left: -50%;
+  text-align: center;
+  position: absolute;
+  vertical-align: middle;
+  left: 0;
+  -webkit-transition: left 500ms, opacity 500ms;
+  -moz-transition: left 500ms, opacity 500ms;
+  -ms-transition: left 500ms, opacity 500ms;
+  -o-transition: left 500ms, opacity 500ms;
+  transition: left 500ms, opacity 500ms;
+}
+
+.carouselPhoto img {
+  position: absolute;
+  left: -9999px;
+  top: -9999px;
+  right: -9999px;
+  bottom: -9999px;
+  margin: auto;
+  overflow: visible;
+}
