+++ client/container/ProductDetail/ProductDetail.css
@@ -1,0 +1,4 @@
+.apd_header {
+  display: flex;
+  margin: 0 14px;
+}
