+++ client/index.css
@@ -2,4 +2,8 @@ body {
   margin: 0;
   padding: 0;
   font-family: sans-serif;
+  font-size: medium;
+  color: #2c2c2c;
+  background-color: #ffffff;
+  display: block;
 }
