+++ public/1.js
@@ -1,2 +1,2 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[1],{52:function(e,n,t){var r=t(53);"string"==typeof r&&(r=[[e.i,r,""]]);var o={hmr:!0,transform:void 0,insertInto:void 0};t(10)(r,o);r.locals&&(e.exports=r.locals)},53:function(e,n,t){(n=e.exports=t(9)(!1)).push([e.i,".client-components-PizzaImage-PizzaImage__PizzaImage--1guqy {\r\n  width: 80%;\r\n  height: 300px;\r\n  margin: 20px auto;\r\n  text-align: center;\r\n}\r\n\r\n.client-components-PizzaImage-PizzaImage__PizzaImg--12yiz {\r\n  max-width: 100%;\r\n  max-height: 100%;\r\n}\r\n",""]),n.locals={PizzaImage:"client-components-PizzaImage-PizzaImage__PizzaImage--1guqy",PizzaImg:"client-components-PizzaImage-PizzaImage__PizzaImg--12yiz"}},54:function(e,n,t){e.exports=t.p+"images/pizzaImage.jpg"},57:function(e,n,t){"use strict";t.r(n);var r=t(0),o=t.n(r),a=t(52),i=t.n(a),c=t(54),u=t.n(c),l=function(e){return o.a.createElement("div",{className:i.a.PizzaImage},o.a.createElement("img",{src:u.a,className:i.a.PizzaImg}))};function f(e){return(f="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function p(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function m(e,n){return!n||"object"!==f(n)&&"function"!=typeof n?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):n}function s(e){return(s=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function z(e,n){return(z=Object.setPrototypeOf||function(e,n){return e.__proto__=n,e})(e,n)}var y=function(e){function n(){return function(e,n){if(!(e instanceof n))throw new TypeError("Cannot call a class as a function")}(this,n),m(this,s(n).apply(this,arguments))}var t,a,i;return function(e,n){if("function"!=typeof n&&null!==n)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(n&&n.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),n&&z(e,n)}(n,r["Component"]),t=n,(a=[{key:"render",value:function(){return o.a.createElement("div",null,o.a.createElement("h1",null,"The Pizza"),o.a.createElement(l,null))}}])&&p(t.prototype,a),i&&p(t,i),n}();n.default=y}}]);
+(window.webpackJsonp=window.webpackJsonp||[]).push([[1],{61:function(e,t,n){var r=n(62);"string"==typeof r&&(r=[[e.i,r,""]]);var o={hmr:!0,transform:void 0,insertInto:void 0};n(7)(r,o);r.locals&&(e.exports=r.locals)},62:function(e,t,n){(t=e.exports=n(6)(!1)).push([e.i,".PizzaImage_1guqy {\r\n  width: 80%;\r\n  height: 300px;\r\n  margin: 20px auto;\r\n  text-align: center;\r\n}\r\n\r\n.PizzaImg_12yiz {\r\n  max-width: 100%;\r\n  max-height: 100%;\r\n}\r\n",""]),t.locals={PizzaImage:"PizzaImage_1guqy",PizzaImg:"PizzaImg_12yiz"}},63:function(e,t,n){e.exports=n.p+"images/pizzaImage.jpg"},66:function(e,t,n){"use strict";n.r(t);var r=n(0),o=n.n(r),i=n(61),a=n.n(i),u=n(63),c=n.n(u),f=function(e){return o.a.createElement("div",{className:a.a.PizzaImage},o.a.createElement("img",{src:c.a,className:a.a.PizzaImg}))};function l(e){return(l="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function p(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function s(e,t){return!t||"object"!==l(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function m(e){return(m=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function y(e,t){return(y=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var b=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),s(this,m(t).apply(this,arguments))}var n,i,a;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&y(e,t)}(t,r["Component"]),n=t,(i=[{key:"render",value:function(){return o.a.createElement("div",null,o.a.createElement("h1",null,"The Pizza"),o.a.createElement(f,null))}}])&&p(n.prototype,i),a&&p(n,a),t}();t.default=b}}]);
 //# sourceMappingURL=1.js.map
