+++ webpack.prod.config.js
@@ -34,7 +34,7 @@ module.exports = {
             options: {
               importLoaders: 1,
               modules: true,
-              localIdentName: '[path][name]__[local]--[hash:base64:5]'
+              localIdentName: '[local]_[hash:base64:5]'
             }
           },
           {
